# Speedup uses of `object.__setattr__`