--- a/changelog.d/898.change.rst
+++ b/changelog.d/898.change.rst
@@ -1,0 +1,1 @@
+Speedup instantiation of frozen slotted classes.