@hynek 2022-01-03
Hmmm 10% more performance is nice and he solution is simpler too, which is nice!

Calling @Tinche who seems to be the daddy of the old code – any edge cases we're missing here?