@Tinche 2022-01-03
This essentially saves us an attribute lookup at the cost of an additional argument in a couple function calls?

If the tests pass, LGTM.